[CHANGED] rgb <-> hsl 的实现
Former-commit-id: 805d5b94700eb2755365eeb4725cd96b57504d58

+++ drag2d/src/drag2d/common/color_convert.h
@@ -6,8 +6,8 @@
 namespace d2d
 {
 
-Colori hsl2rgb(float h, float s, float l);
-Colori rgb2hsl(float r, float g, float b);
+Colorf hsl2rgb(float h, float s, float l);
+Colorf rgb2hsl(float r, float g, float b);
 
 }
